Creates and finalizes footer: 1- moves side-bar inside of footer. 2- reorganize folders
