--- src/components/main/Footer.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Footer() {
-  return <footer className="app-container">Footer</footer>;
-}
