+++ src/components/main/Footer/Footer.jsx
@@ -1,0 +1,9 @@
+import SideNav from './SideNav';
+
+export default function Footer() {
+  return (
+    <footer>
+      <SideNav />
+    </footer>
+  );
+}
