+++ src/styles/components/header/index.scss
@@ -1,3 +1,3 @@
 @forward 'logo';
+@forward 'header';
 @forward 'mobile-menu';
-@forward 'side-nav';
