+++ src/styles/layouts/_footer.scss
@@ -1,0 +1,56 @@
+@use '../utilities/mixins' as u;
+
+.side-nav {
+  display: flex;
+  position: fixed;
+  bottom: 0;
+  background-color: var(--background-color-light);
+  height: 4em;
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+  box-shadow: 0 0 0.3em var(--drop-shadow-light);
+  z-index: 1;
+  border-top: 1px solid var(--text-color-dark-secondary);
+  ul {
+    padding: 0;
+    margin: 0;
+    display: flex;
+    justify-content: space-between;
+    li {
+      a,
+      button {
+        padding: 1em 1.5em;
+        display: flex;
+        transition: all 0.5s;
+        div {
+          padding: 0 1em;
+        }
+      }
+    }
+  }
+  @include u.breakpoint-min('medium') {
+    left: 0;
+    height: 100%;
+    width: 4.7em;
+    align-items: flex-end;
+    justify-content: center;
+    box-shadow: none;
+    border-top: none;
+    background-color: transparent;
+    ul {
+      flex-direction: column;
+
+      li {
+        display: flex;
+        padding: 1em 0;
+        align-items: center;
+        justify-content: center;
+
+        div {
+          padding: 0 1em;
+        }
+      }
+    }
+  }
+}
